+++ ng-workspace/projects/OpenWaterFoundation/common/ui/dashboard/widget/image/image.component.ts
@@ -15,9 +15,9 @@ export class ImageComponent {
   /** The attribute provided as an attribute to this component when created, e.g.
    *   <widget-image [imageWidget]="path/to/image.png"></widget-image> */
   @Input() imageWidget: IM.ImageWidget;
-  /** The string representing the type of error that occurred while building this
-    * widget. Used by the error widget. */
-  errorType: string;
+  /** String array representing the type of error that occurred while building this
+   * widget. Used by the error widget. */
+  errorTypes: string[] = [];
   /** The path to the image file after it has been built with the OWF service to
    * deal with either an absolute or relative path provided. */
   fullDataPath: string;
@@ -33,17 +33,30 @@ export class ImageComponent {
   constructor(private commonService: OwfCommonService) {}
 
 
+  private checkWidgetObject(): void {
+
+    if (!this.imageWidget.imagePath) {
+      this.widgetError = true;
+      this.errorTypes.push('no imagePath');
+      return;
+    }
+    if (!this.imageWidget.name) {
+      this.widgetError = true;
+      this.errorTypes.push('no name');
+      return;
+    }
+
+    this.fullDataPath = this.commonService.buildPath(IM.Path.dbP, [this.imageWidget.imagePath]);
+  }
+
   /**
    * Called right after the constructor.
    */
   ngOnInit(): void {
 
-    if (!this.imageWidget.imagePath) {
-      this.widgetError = true;
-      this.errorType = 'no imagePath';
-      return;
-    }
-    this.fullDataPath = this.commonService.buildPath(IM.Path.dbP, [this.imageWidget.imagePath]);
+    this.checkWidgetObject();
+
+    
   }